+++ nj-core/src/basic.rs
@@ -174,6 +174,59 @@ impl JsEnv {
     }
 
     #[allow(clippy::not_unsafe_ptr_arg_deref)]
+    pub fn coerce_to_object(&self, value: napi_value) -> Result<napi_value, NjError> {
+        let mut napi_obj = ptr::null_mut();
+        napi_call_result!(crate::sys::napi_coerce_to_object(
+            self.0,
+            value,
+            &mut napi_obj
+        ))?;
+        Ok(napi_obj)
+    }
+
+    #[allow(clippy::not_unsafe_ptr_arg_deref)]
+    pub fn get_property_names(&self, js_obj: napi_value) -> Result<napi_value, NjError> {
+        let mut js_arr = ptr::null_mut();
+        napi_call_result!(crate::sys::napi_get_property_names(
+            self.0,
+            js_obj,
+            &mut js_arr
+        ))?;
+        Ok(js_arr)
+    }
+
+    #[allow(clippy::not_unsafe_ptr_arg_deref)]
+    pub fn get_property(
+        &self,
+        js_obj: napi_value,
+        name: napi_value,
+    ) -> Result<napi_value, NjError> {
+        let mut value = ptr::null_mut();
+        napi_call_result!(crate::sys::napi_get_property(
+            self.0, js_obj, name, &mut value
+        ))?;
+        Ok(value)
+    }
+
+    #[allow(clippy::not_unsafe_ptr_arg_deref)]
+    pub fn set_property(
+        &self,
+        object: napi_value,
+        name: &str,
+        value: napi_value,
+    ) -> Result<(), NjError> {
+        // napi does not accept a length argument here, so name MUST be nul-terminated
+        let name_utf8 = [name.as_bytes(), &[0x00]].concat();
+        napi_call_result!(crate::sys::napi_set_named_property(
+            self.0,
+            object,
+            name_utf8.as_ptr() as *const ::std::os::raw::c_char,
+            value
+        ))?;
+        Ok(())
+    }
+
+    #[allow(clippy::not_unsafe_ptr_arg_deref)]
     pub fn get_element(&self, array: napi_value, index: u32) -> Result<napi_value, NjError> {
         let mut element = ptr::null_mut();